The page says which one thing to do next
The queue has answered "what matters most" since the ranking shipped — it
is the first row. What it never did is SAY so. A reader arriving at
twenty-five rows has to work out that the top one is the answer, and a
list is a shape that invites scanning rather than acting.

So the first row is lifted into a card with the evidence a rep needs to
act without opening anything: what it is, why now in money and dates,
what it costs to do nothing, and the ONE verb the server already names in
primary_action. The comparator stays off it — "182 against 180" belongs
behind a disclosure, never in the primary scan path.

The row stays in the queue below the card. Removing it would make the
rank numbers lie and the counts disagree with the page.

NO CARD OVER REVIEW WORK. A day whose top row is a duplicate-merge
suggestion is not a day with a recommended action, and a page headed "do
this next" over one would tell a rep something false about their morning.
No card either where the server named no verb: a card whose button the
page has to invent is a card that guesses.

worklist.tsx was 588 lines against a 500 cap, so the row and its verbs
move to worklist.row.tsx — 294 and 316. They answer different questions:
the screen decides what the page shows, a row decides how one piece of
work reads and where each verb goes.

Three tests, the exclusion mutation-checked. Verified against the running
product: the top row is a task in keep_momentum whose verb is `complete`,
and the card draws.
