+++ frontend/src/i18n/de.ts
@@ -7360,6 +7360,16 @@ export const de = {
     "Unter den Quellen, die geantwortet haben, wartet nichts.",
   "worklist.partial": "{sources} — das ist nicht der ganze Tag.",
   "worklist.overdue": "Überfällig",
+  "worklist.focus.title": "Das als Nächstes",
+  "worklist.focus.verb.decide": "Entscheiden",
+  "worklist.focus.verb.merge": "Paar prüfen",
+  "worklist.focus.verb.complete": "Erledigen",
+  "worklist.focus.verb.act": "Bearbeiten",
+  "worklist.focus.verb.acknowledge": "Zur Kenntnis nehmen",
+  "worklist.focus.verb.open": "Öffnen",
+  "worklist.focus.verb.snooze": "Öffnen",
+  "worklist.focus.verb.dismiss": "Öffnen",
+  "worklist.focus.verb.set_aside": "Öffnen",
   "worklist.band.now": "Jetzt",
   "worklist.band.build_pipeline": "Pipeline aufbauen",
   "worklist.band.keep_momentum": "In Bewegung halten",
