+++ frontend/src/i18n/en.ts
@@ -7430,6 +7430,16 @@ export const en = {
     "Nothing is waiting among the sources that answered.",
   "worklist.partial": "{sources}, so this is not the whole day.",
   "worklist.overdue": "Overdue",
+  "worklist.focus.title": "Do this next",
+  "worklist.focus.verb.decide": "Decide",
+  "worklist.focus.verb.merge": "Review the pair",
+  "worklist.focus.verb.complete": "Complete it",
+  "worklist.focus.verb.act": "Act on it",
+  "worklist.focus.verb.acknowledge": "Acknowledge",
+  "worklist.focus.verb.open": "Open it",
+  "worklist.focus.verb.snooze": "Open it",
+  "worklist.focus.verb.dismiss": "Open it",
+  "worklist.focus.verb.set_aside": "Open it",
   "worklist.band.now": "Now",
   "worklist.band.build_pipeline": "Build pipeline",
   "worklist.band.keep_momentum": "Keep momentum",
