+++ frontend/src/i18n/vi.ts
@@ -7284,6 +7284,16 @@ export const vi = {
     "Không có gì đang chờ trong các nguồn đã trả lời.",
   "worklist.partial": "{sources}, nên đây chưa phải cả ngày.",
   "worklist.overdue": "Quá hạn",
+  "worklist.focus.title": "Làm việc này tiếp theo",
+  "worklist.focus.verb.decide": "Quyết định",
+  "worklist.focus.verb.merge": "Xem cặp trùng",
+  "worklist.focus.verb.complete": "Hoàn thành",
+  "worklist.focus.verb.act": "Xử lý",
+  "worklist.focus.verb.acknowledge": "Đã biết",
+  "worklist.focus.verb.open": "Mở",
+  "worklist.focus.verb.snooze": "Mở",
+  "worklist.focus.verb.dismiss": "Mở",
+  "worklist.focus.verb.set_aside": "Mở",
   "worklist.band.now": "Bây giờ",
   "worklist.band.build_pipeline": "Xây dựng pipeline",
   "worklist.band.keep_momentum": "Giữ đà",
